Don't import from conftest
Looks like this might have been missed in the original PR.
It's unusual to import explicitly from conftest modules.

Additionally, now that the package tests have moved into the
functional/ tests, the only shared bits of code was the
FakePipCall, which I've just defined in both modules.
Seems like a reasonable compromise to me, it's a one line
namedtuple definition.
